+++ ckanext/gobar_theme/helpers/helpers.py
@@ -12,14 +12,3 @@ def attributes_has_at_least_one(attr, resource_attributes):
         if attributes.get(attr, ''):
             return True
     return False
-
-
-def portal_andino_version():
-    from ckanext.gobar_theme.actions import _get_portal_andino_version
-    version = _get_portal_andino_version()
-    version = version['portal-andino'] or 'Desarrollo'
-
-    version = version.replace('release-', '')  # Elimino el release-
-    version = version[:15]  # me quedo con los primeros 15 caracteres
-
-    return version
